# internal/swagger/swagger.yaml
basePath: /api
definitions:
...
  github_com_neu-software-practice_software-practice-backend_internal_dto.RegisterRequest:
    properties:
      age:
        type: integer
      age_type:
        type: string
      birthdate:
        description: YYYY-MM-DD
        type: string
      card_number:
        type: string
      deptment_id:
        type: integer
      employee_id:
        type: integer
      gender:
        type: string
      home_address:
        type: string
      is_book:
        type: string
      real_name:
        type: string
      regist_level_id:
        type: integer
      regist_method:
        type: string
      settle_category_id:
        type: integer
    required:
    - deptment_id
    - employee_id
    - gender
    - real_name
    - regist_level_id
    - settle_category_id
    type: object
...
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 门诊全流程后端：挂号收费 / 门诊医生 / 检查 / 检验 / 处置 / 药房。
  title: 东软云医院 HIS 门诊管理 API
  version: "1.0"
paths:
...
  /registers:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        type: string
      - description: 姓名
        in: query
        name: name
        type: string
      - description: 看诊状态(1已挂号/2接诊/3结束/4退号)
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 挂号记录查询 (F1-2)
      tags:
      - registration
    post:
      consumes:
      - application/json
      parameters:
      - description: 挂号信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 窗口挂号
      tags:
      - registration
  /registers/{id}/cancel:
    post:
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 窗口退号 (F1-2)
      tags:
      - registration
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"